Adds auctions list to favourites page
The favourites page now lists auctions that are being watched by the currently logged in user. As this is user specific data, we retrieve auction data on client-side.

This also adds realtime updates which will trigger a data fetch if the user adds/removes a favourite in a separate tab.